refactor: Promote handling part of ScoreboardModel to ScoreboardHandler
