+++ common/src/main/java/com/wynntils/handlers/scoreboard/ScoreboardLine.java
@@ -2,6 +2,6 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.wynn.model.scoreboard;
+package com.wynntils.handlers.scoreboard;
 
 public record ScoreboardLine(String line, int index) {}
