+++ common/src/main/java/com/wynntils/handlers/scoreboard/ScoreboardLineChange.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.wynn.model.scoreboard;
+package com.wynntils.handlers.scoreboard;
 
 import net.minecraft.server.ServerScoreboard;
