+++ common/src/main/java/com/wynntils/handlers/scoreboard/Segment.java
@@ -2,8 +2,9 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.wynn.model.scoreboard;
+package com.wynntils.handlers.scoreboard;
 
+import com.wynntils.wynn.model.scoreboard.ScoreboardModel;
 import java.util.ArrayList;
 import java.util.List;
