+++ common/src/main/java/com/wynntils/wynn/model/GuildAttackTimerModel.java
@@ -9,8 +9,8 @@ import com.wynntils.core.managers.Model;
 import com.wynntils.utils.Pair;
 import com.wynntils.utils.objects.TimedSet;
 import com.wynntils.wynn.event.ChatMessageReceivedEvent;
-import com.wynntils.wynn.model.scoreboard.ScoreboardListener;
-import com.wynntils.wynn.model.scoreboard.Segment;
+import com.wynntils.handlers.scoreboard.ScoreboardListener;
+import com.wynntils.handlers.scoreboard.Segment;
 import com.wynntils.wynn.model.scoreboard.guild.GuildAttackListener;
 import com.wynntils.wynn.model.scoreboard.guild.TerritoryAttackTimer;
 import java.util.ArrayList;
