+++ common/src/main/java/com/wynntils/wynn/model/quests/QuestManager.java
@@ -15,7 +15,7 @@ import com.wynntils.mc.utils.McUtils;
 import com.wynntils.wynn.event.QuestBookReloadedEvent;
 import com.wynntils.wynn.event.TrackedQuestUpdateEvent;
 import com.wynntils.wynn.event.WorldStateEvent;
-import com.wynntils.wynn.model.scoreboard.ScoreboardListener;
+import com.wynntils.handlers.scoreboard.ScoreboardListener;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
