+++ common/src/main/java/com/wynntils/wynn/model/quests/QuestScoreboardListener.java
@@ -7,9 +7,9 @@ package com.wynntils.wynn.model.quests;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.managers.Managers;
 import com.wynntils.mc.utils.ComponentUtils;
-import com.wynntils.wynn.model.scoreboard.ScoreboardListener;
+import com.wynntils.handlers.scoreboard.ScoreboardListener;
+import com.wynntils.handlers.scoreboard.Segment;
 import com.wynntils.wynn.model.scoreboard.ScoreboardModel;
-import com.wynntils.wynn.model.scoreboard.Segment;
 import com.wynntils.wynn.utils.WynnUtils;
 import java.util.List;
 import net.minecraft.ChatFormatting;
